+++ src/components/layouts/Laptop/LaptopBodyWall.jsx
@@ -1,12 +1,12 @@
 import React, {Component} from 'react';
 import {Card, Col, Container, Row} from "react-bootstrap";
 
-import LaptopWall from './../../../assets/images/Laptop/LaptopWall.jpg';
+import LaptopWall from '../../../assets/images/Laptop/LaptopWall.jpg';
 
-class LaptopImageUploadBodyWall extends Component {
+class LaptopBodyWall extends Component {
 
     divBack = {
-        backgroundColor : '#212121',
+        'background-color' : '#212121',
         color : 'white'
     }
 
@@ -21,7 +21,7 @@ class LaptopImageUploadBodyWall extends Component {
                 <Container style={this.divBack}>
                     <Row>
                         <Col sm={4}>
-                            <h1>Laptop Image Upload</h1>
+                            <h1>Laptops</h1>
                         </Col>
                         <Col sm={8}>
                             <Card>
@@ -38,4 +38,4 @@ class LaptopImageUploadBodyWall extends Component {
     }
 }
 
-export default LaptopImageUploadBodyWall;
+export default LaptopBodyWall;
