+++ src/components/layouts/Laptop/LaptopImageUploadBodyWall.jsx
@@ -1,9 +1,9 @@
 import React, {Component} from 'react';
-import {Card, Col, Container, Image, Row} from "react-bootstrap";
+import {Card, Col, Container, Row} from "react-bootstrap";
 
-import LaptopWall from './../../../assets/images/Laptop/LaptopWall.jpg';
+import LaptopWall from '../../../assets/images/Laptop/LaptopWall.jpg';
 
-class LaptopDashboardBodyWall extends Component {
+class LaptopImageUploadBodyWall extends Component {
 
     divBack = {
         backgroundColor : '#212121',
@@ -21,7 +21,7 @@ class LaptopDashboardBodyWall extends Component {
                 <Container style={this.divBack}>
                     <Row>
                         <Col sm={4}>
-                            <h1>All Laptop</h1>
+                            <h1>Laptop Image Upload</h1>
                         </Col>
                         <Col sm={8}>
                             <Card>
@@ -38,4 +38,4 @@ class LaptopDashboardBodyWall extends Component {
     }
 }
 
-export default LaptopDashboardBodyWall;
+export default LaptopImageUploadBodyWall;
