+++ src/components/pages/Laptop.jsx
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import LaptopBodyWall from "../sections/Laptop/LaptopBodyWall";
+import LaptopBodyWall from "../layouts/Laptop/LaptopBodyWall";
 import LaptopListHomepage from "../sections/Laptop/LaptopListHomepage";
 
 class Laptop extends Component {
