+++ src/components/sections/Laptop/LaptopImageUpload.jsx
@@ -4,7 +4,7 @@ import {Button, Container, Form, ProgressBar} from "react-bootstrap";
 import {storage} from '../../../firebase/FirebaseLaptop';
 import ServiceLaptopImage from "../../../services/ServiceLaptopImage";
 import NavigationBarDashboard from "../../layouts/Navigation/NavigationBarDashboard";
-import LaptopImageUploadBodyWall from "./LaptopImageUploadBodyWall";
+import LaptopImageUploadBodyWall from "../../layouts/Laptop/LaptopImageUploadBodyWall";
 import ServiceLaptop from "../../../services/ServiceLaptop";
 
 function LaptopImageUpload() {
